Add more precise overloads for `agg`
`DataFrame.agg` and `Series.agg` have multiple possible return types
per https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.agg.html#pandas.DataFrame.agg
and https://pandas.pydata.org/docs/reference/api/pandas.Series.agg.html#pandas.Series.agg;
note that the possible return types are different between `DataFrame.agg` and `Series.agg`,
and also depend on the type of the `func` input argument.

Add more precise overloads for the cases that mypy can statically handle:
- multiple aggregation functions
  (either a list/dict, OR if `func=None` where the funcs are specified via `kwargs`)
- the `str` and `Callable` cases for a single aggregation function

numpy may not be in the mypy venv,
in which case numpy functions would be seen as `Any`
and thus can't be used for `@overload` type dispatching in mypy,
so don't be precise in the case of a numpy function.
When mypy sees it as `Any`, trying to be precise would cause errors like
`Overloaded function signatures 1 and 2 overlap with incompatible return types`.

Additionally, avoid `Scalar` as a return type even when it would be correct.
`Scalar` is a large Union that produces many (currently 13) lines of dense error output
when there is a type mismatch;
given users would almost always have to `cast` in this case,
use `Any` instead and avoid providing a type hint in this case.

The docs recommend using the `.agg` alias for `.aggregate`,
so don't provide overloads for the latter to nudge folks towards the former.

Add test cases (from the docs) for lists and dicts of aggregation functions.